Add history order page & use blueprint
